Bueno ahi esta la logica basica completa, faltaria hacer solamente las otras personalidades, solo esta la del soltero. Y despues el tema de dibujar la ciudad faltaria tambien

+++ celdas-popper/src/main/java/ar/uba/fi/celdas7568/ciudad/Personalidad.java
@@ -2,22 +2,22 @@ package ar.uba.fi.celdas7568.ciudad;
 
 /**
  * Cada valor tiene un rango entre -1 y 1 que indica la importancia que la persona le da a la caracteristica.
- * Valores negativos indican que la caracter’stica es indeseable
- * Valores positivos indican que la caracter’stica es deseable
- * Valores cero indican que la caracter’stica es indistinta
+ * Valores negativos indican que la caracterï¿½stica es indeseable
+ * Valores positivos indican que la caracterï¿½stica es deseable
+ * Valores cero indican que la caracterï¿½stica es indistinta
  * @author emmanuelespina
  *
  */
 public class Personalidad {
-	public float cultura;
-	public float diversion;
-	public float seguridad;
-	public float educacion;
-	public float naturaleza;
-	public float populoso;
-	public float familiar;
-	public float tranquilo;
-	public float transporte;
-	public float barrioExclusivo;
-	public float costo;
+	public double cultura;
+	public double diversion;
+	public double seguridad;
+	public double educacion;
+	public double naturaleza;
+	public double populoso;
+	public double familiar;
+	public double tranquilo;
+	public double transporte;
+	public double barrioExclusivo;
+	public double costo;
 }
